Add include path resolving callback.
By default, the preprocessor uses raw file name as includeOpen argument
and for internal storage (i.e. for errors). While it is possible to
resolve the file path using the parent path in includeOpen, this requires
keeping a mapping of source -> file path strings, and does not solve
errors or other parts where filename might be used.

Provide an optional callback that can resolve include paths. If the
callback does not exist, the behavior is the same as before.